rename and centralisation of data model
