+++ src-tauri/src/commands/file_system_operation_commands.rs
