+++ src-tauri/src/commands/mod.rs
@@ -1,4 +1,4 @@
 pub mod meta_data_commands;
 mod fs_dir_loader_commands;
 mod selected_path_for_action_commands;
-pub mod file_operations;
+pub mod file_system_operation_commands;
