+++ src-tauri/src/filesystem/models/directory_entries.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+use crate::filesystem::models;
+
+#[derive(Serialize, Deserialize, Debug, Clone, Eq, PartialEq, Hash)]
+pub struct Entries {
+    pub(crate) directories: Vec<models::Directory>,
+    pub(crate) files: Vec<models::File>,
+}
