+++ src-tauri/src/filesystem/models/mod.rs
@@ -4,4 +4,5 @@ mod file;
 pub use file::File;
 mod volume;
 pub use volume::VolumeInformation;
-
+mod directory_entries;
+pub use directory_entries::Entries;
